Change op-amp instantiation
Op-amps were previously defined using their library name, and parameters
were pulled from the shared library file. That prevented new op-amps from
being added dynamically in one module and used in another (for example
when the LISO module parses an output file's op-amp - creating it in the
library - it is not available to the circuit module). Now op-amps are
instantiated with their parameters, and the op-amp library simply provides
a factory method to get op-amp objects by their model number.