Compute noise on a per-source basis
Previously, the solver computed noise at a node by computing the noise entering at
each node or component and projecting it to that node. This worked except for
nodes with multiple current noise sources (e.g. two op-amps connected to the same
node). It also made it difficult to properly label a noise source, as the noise
could contain contributions from multiple sources.

Now the solver computes noise on a per-source basis such that the noise arising
from a single source is computed to the output and able to be plotted separately.
The solution object will in the future support some ability to combine sources
of noise at the same node together, to support the previous behaviour.