Revamp circuit solver to properly solve noise and transfer functions
The Input() component now provides the necessary coefficients for the circuit matrix
in order to accurately simulate noise and voltage/current transfer functions.

The Circuit() object also now takes minimal inputs for its solve method, which has
itself been removed and split into two: calculate_tfs() and calculate_noise().

As the circuit matrix and equation printers required the circuit matrix, and the
circuit matrix was different for noise and transfer function calculations, now
these methods are private and must be provided with a matrix as an argument. In
order to print the matrix/equations, now parameters must be set in the call to
Circuit.calculate_tfs() or Circuit.calculate_noise().

The LISO parser has been modified to work with the new Circuit interface. The tests
comparing native simulations to those of the LISO binary still pass.