add graphviz visual output
This adds a node_graph() function to Circuit, which returns a graphviz
object containing a visual representation of the circuit. For jupyter notebook
environments, it also provides a _repr_svg_ function for Circuit.

Needs the graphviz python package.
Some call to the node_graph() function should probably be provided via the
Parser as well. Provide convenience binding that immediately produces a .pdf
output with each simulation run?