Throw custom exception within input parser
The parser previously threw SyntaxError when encountering a parsing error not caught by PLY, such as
"freq dec 1 1e4 101" ("dec" is invalid). These were caught by PLY and not rethrown, but instead added
to the parsing stack for another rule to handle. Instead of handling the error with another rule, the
parser now throws a custom LisoParserError that is not caught by PLY. This immediately terminates
execution when invalid syntax is provided.