Fix unit parsing in Input component
Previously, the Input component parsed its input impedance as a float, meaning that strings like "10M" were not
supported. Furthermore, the input parser did not catch this error because it did not parse the input until the
circuit was being built.

Now the quantity is properly parsed and there is a new test to check this.