Add experimental branch with compatibility to LISO for voltage followers
`circuit` does not simulate voltage followers the same way as LISO does without this patch. It sets the inverse voltage
gain of the output of a voltage follower (buffer) op-amp to 1 + 1/H(s) instead of 1/H(s), which makes it agree more
closely with LISO; however, I am not sure if this behaviour is correct in LISO - the output of the "voltage-follower"
transfer function test should I believe be unity within the gain bandwidth of the op-amp, but instead it's almost zero.