Change matching methods for data sets
The __eq__ methods in data sets were used for checking if data set series were closely matched.
Since data sets must be checked for equality when merging solutions, but in this case in terms
of their meta data, the __eq__ method seemed more suitable for that purpose, and so the method
used to check for data set equality was changed to "matches". At the same time, some
restructuring of the hierarchy was made.