Change solution merging behaviour
Groups were previously prepended with the solution name to avoid conflicts. This caused lots of extra
complexity in other code, and created some strange labels when solutions were merged multiple times.
Now group merging is entirely "dumb": groups with the same name get merged in the resulting solution.
The rule regarding identical functions still stands, so identical functions in identical groups will
throw an error if attempted to be merged. It is now up to the user to rename groups appropriately if
they wish to keep groups with the same name separate after a merge (this is done in the CLI).